Add `<Menu.ResourceItem>`component to facilitate menu customization
